# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PORTAL API Documentation",
    "version": "1.0.0",
    "description": "Documentación técnica de la API de PORTAL - Sistema de Gestión de Personal y Horarios."
  },
  "servers": [{ "url": "http://localhost:4000", "description": "Servidor de Desarrollo" }],
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": { "success": { "type": "boolean" }, "message": { "type": "string" } },
        "required": ["success"]
      },
...
      "Employee": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "rut": { "type": "string", "minLength": 8 },
          "name": { "type": "string", "minLength": 2 },
          "position": { "type": "string" },
          "department": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string" },
          "emergencyContact": { "type": "string" },
          "status": { "type": "string", "enum": ["Activo", "Archivado"] },
          "contractType": { "type": "string" },
          "shiftPatternId": { "type": "string" },
          "hireDate": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
          "terminationDate": { "type": "string", "nullable": true },
          "address": { "type": "string" },
          "commune": { "type": "string" },
          "region": { "type": "string" },
          "birthDate": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
          "gender": { "type": "string" },
          "bankName": { "type": "string" },
          "accountType": { "type": "string" },
          "accountNumber": { "type": "string" },
          "createUserAccount": { "type": "boolean" },
          "isDeleted": { "type": "boolean" },
          "syncStatus": { "type": "string" },
          "lastModified": { "type": "number" }
        },
        "required": ["rut", "name"]
      },
...
      "MeterReading": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "meterConfigId": { "type": "string" },
          "authorUsername": { "type": "string" },
          "value": { "type": "number" },
          "timestamp": { "type": "string" },
          "isRecharge": { "type": "boolean" },
          "notes": { "type": "string", "nullable": true },
          "createdAt": { "type": "string" },
          "isDeleted": { "type": "boolean" },
          "syncStatus": { "type": "string" },
          "lastModified": { "type": "number" }
        },
        "required": ["meterConfigId", "authorUsername", "value"]
      },
      "MeterReadingListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": { "type": "array", "items": { "$ref": "#/components/schemas/MeterReading" } }
            },
            "required": ["data"]
          }
        ]
      },
...
      "PaginatedMeterReadingResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "pagination": {
                "type": "object",
                "properties": {
                  "total": { "type": "number" },
                  "page": { "type": "number" },
                  "totalPages": { "type": "number" }
                },
                "required": ["total", "page", "totalPages"]
              },
              "data": { "type": "array", "items": { "$ref": "#/components/schemas/MeterReading" } }
            },
            "required": ["pagination", "data"]
          }
        ]
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/meters": {
      "get": {
        "summary": "Obtiene lecturas de medidores (soporta paginación)",
        "tags": ["Meters"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer" } },
          { "in": "query", "name": "pageSize", "schema": { "type": "integer" } },
          { "in": "query", "name": "meterId", "schema": { "type": "string" } },
          { "in": "query", "name": "startDate", "schema": { "type": "string", "format": "date" } },
          { "in": "query", "name": "endDate", "schema": { "type": "string", "format": "date" } },
          {
            "in": "query",
            "name": "since",
            "schema": { "type": "string" },
            "description": "Timestamp para consulta delta"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de lecturas",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/PaginatedMeterReadingResponse" },
                    { "$ref": "#/components/schemas/MeterReadingListResponse" }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/meters/bulk": {
      "post": {
        "summary": "Registra lecturas de medidores en lote",
        "tags": ["Meters"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/MeterReading" }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lecturas registradas",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeterReadingListResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/import/preview": {
      "post": {
        "summary": "Previsualizar contenido de un archivo Excel",
        "tags": ["Import"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" },
                  "schema": { "type": "string", "description": "JSON string of the mapping schema" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Datos extraídos del archivo" } }
      }
    },
...
    "/api/export/shift-report-pdf/{id}": {
      "get": {
        "summary": "Exportar bitácora de turno a PDF",
        "tags": ["Exports"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Archivo PDF",
            "content": { "application/pdf": { "schema": { "type": "string", "format": "binary" } } }
          }
        }
      }
    },
    "/api/export/report-excel": {
      "get": {
        "summary": "Exportar reporte de asistencia a Excel",
        "tags": ["Exports"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": { "type": "string", "format": "date" },
            "required": true
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": { "type": "string", "format": "date" },
            "required": true
          },
          { "in": "query", "name": "area", "schema": { "type": "string" } },
          { "in": "query", "name": "employeeId", "schema": { "type": "string" } },
          {
            "in": "query",
            "name": "mode",
            "schema": { "type": "string", "enum": ["summary", "detailed"] }
          }
        ],
        "responses": { "200": { "description": "Archivo Excel" } }
      }
    },
...
    "/api/employees": {
      "get": {
        "summary": "Obtiene todos los empleados",
        "tags": ["Employees"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer" } },
          { "in": "query", "name": "pageSize", "schema": { "type": "integer" } },
          { "in": "query", "name": "search", "schema": { "type": "string" } },
          { "in": "query", "name": "status", "schema": { "type": "string" } },
          { "in": "query", "name": "area", "schema": { "type": "string" } },
          { "in": "query", "name": "since", "schema": { "type": "string" } }
        ]
      },
      "post": {
        "summary": "Crea un nuevo empleado",
        "tags": ["Employees"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Employee" } }
          }
        },
        "responses": {
          "201": {
            "description": "Empleado creado",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Employee" } }
            }
          }
        }
      }
    },
    "/api/employees/bulk": {
      "post": {
        "summary": "Importación masiva de empleados",
        "tags": ["Employees"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Employee" } }
            }
          }
        },
        "responses": { "201": { "description": "Empleados creados exitosamente" } }
      }
    },
    "/api/employees/{id}": {
      "put": {
        "summary": "Actualiza un empleado existente",
        "tags": ["Employees"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Employee" } }
          }
        },
        "responses": {
          "200": {
            "description": "Empleado actualizado",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Employee" } }
            }
          }
        }
      }
    },
    "/api/employees/export": {
      "get": {
        "summary": "Exportar empleados en formato Excel/CSV",
        "tags": ["Employees"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": { "type": "string", "enum": ["Activo", "Archivado"] }
          },
          { "in": "query", "name": "search", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Archivo generado",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          }
        }
      }
    },
...
  },
...
}